chore(duplication): share vscode metadata helpers (#5471)
